fixed with del rooma and i start to work with lab

+++ app/src/main/java/com/example/recyclerviewappintent/TodoEntity.kt
@@ -6,6 +6,6 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "todos")
 class TodoEntity {
     @PrimaryKey(autoGenerate = true)
-    var id: Long? = null
+    var id: Int? = null
     var content: String = ""
 }
